iommufd: Add device_users to track the hw_pagetable usage by device
Currently, hw_pagetable tracks the attached devices in the device list.
When attaching to the kernel-managed hw_pagetable, it should be linked
to IOAS if it's the first device attached to ths hw_pagetable, and
disconnected from IOAS/IOPT if it is the last device. This is checked by
list_empty(hwpt->devices) as all hw_pagetable tracks attached devices
in this list.

However, under nested translation, device is attached to the user-managed
stage-1 hw_pagetable, it won't be added to the kernel-managed stage-2
hw_pagetable's device list. This breaks the logic for the kernel-managed
hw_pagetable link/disconnect to/from IOAS/IOPT. e.g. the stage-2 hw_pagetable
would be linked to IOAS multiple times if multiple device is attached,
but it will become empty as soon as one device detached.

This adds device_users in struct iommufd_hw_pagetable to track the users
of hw_pagetable. This field is a pointer, only do memory allocation for the
stage-2 hw_pagetable. For stage-1 hw_pagetable, it should reuse the stage-2
hw_pagetable's device_user. When device is attached to hw_pagetable, iommufd
increases the device_users no matter it is stage-1 or stage-2 hw_pagetable.

Adding this device_users also reduces the dependency on hwpt->devices list,
so it helps the following patch to remove the device list completely.

Co-developed-by: Yi Liu <yi.l.liu@intel.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>