iommufd/hw_pagetable: Add infrastruture for nested translation
Under nested translation, device is attached to the stage-1 hw_pagetable,
while the stage-2 hw_pageable is a background object. This is done by
storing a parent hw_pagetable in stage-1 hw_pagetable. When doing device
attachment,
  1) device is firstly "attached" to the stage-2 hw_pagetable. This
     attachment sets up the stage-2 page table e.g. add to IOAS/IOPT.
     And increase the device_users.

  2) device is added to the stage-1 hw_pagetable's device_list.

For the stage-1 hw_pagetable in nested, it is maintained by userspace, so
it requires userspace to provide data for creation. This will be added in
further commits.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>