cover-letter: Add IO page table replacement support
Hi all,

======== internal review ========

Changelog:
v4->v5:
 * Rebased the change on top of two being-reviewed series
   a) broken_unmanaged_domain b) remove iommufd_hw_pagetable_has_group
 * Added Yi's change of introducing iommu-priv.h header
 * Moved the iommu_group_replace_domain() prototype to iommu-priv.h
 * Dropped "group->domain == new_domain" comparison since it's included
   in the __iommu_group_set_domain().
 * Fixed the scope of access->ioas_lock.
 * Dropped related-hwpt check before calling iommufd_device_attach_ioas.
 * Updated commit messages by incorporating comments from Jason/Robin.
v3->v4:
 * Renamed "no_iommufd" to "broken_unmanaged_domain".
 * Renamed "iommu_replace_group" to "iommu_group_replace_domain".
 * Rebased on top of Jason's "iommu: Reorganize __iommu_attach_group()".
 * Fixed uninitialized "ret" bug in iommu_group_replace_domain().
 * Changed iommu_group_replace_domain's export symbol to IOMMUFD_INTERNAL.
 * Added fallback routine in iommu_group_replace_domain() upon failures
   of __iommu_group_set_domain() calls.
 * Added a patch to block broken_unmanaged_domain in iommufd_device_bind
 * Added a patch to drop iommufd_hw_pagetable_has_group(). This removes
   the device list and devices_lock, so a later patch can be simplified.
 * Added an access->ioas_unpin for unpin_pages() to use exclusively, so
   there's no longer a need of a manual unpin_pages_all() function.
 * Updated kdoc and commit messages accordingly.
v2->v3:
 * Added a no_iommufd flag to older drivers
 * Fail iommu_replace_group() for drivers with no_iommufd=true
 * Reworked the access->ioas replacement
 * Added selftest coverage for access->ioas replacement
v1->v2:
 * Rebased on top of Yi's "wip/iommufd-v6.2-rc4-nesting" branch
 * Rebased on top of Baolu's "Retire detach_dev callback" series
 * Added a mutex to protect access->ioas
 * Dropped domain_replace() op, by using attach_dev() directly.
 * Renamed to iommu_replace_group and moved it next to iommu_attach_group

======== cover-letter ========

The existing IOMMU APIs provide a pair of functions: iommu_attach_group()
for callers to attach a device from the default_domain (NULL if not being
supported) to a given iommu domain, and iommu_detach_group() for callers
to detach a device from a given domain to the default_domain. Internally,
the detach_dev op is deprecated for the newer drivers with default_domain.
This means that those drivers likely can switch an attaching domain to
another one, without stagging the device at a blocking or default domain,
for use cases such as:
1) vPASID mode, when a guest wants to replace a single pasid (PASID=0)
   table with a larger table (PASID=N)
2) Nesting mode, when switching the attaching device from an S2 domain
   to an S1 domain, or when switching between relevant S1 domains.

This series introduces a new iommu_group_replace_domain() for that. And
add corresponding support throughout the uAPI. So user space can do such
a REPLACE ioctl reusing the existing VFIO_DEVICE_ATTACH_IOMMUFD_PT. This
means that user space needs to be aware whether the device is attached or
not: an unattached device calling VFIO_DEVICE_ATTACH_IOMMUFD_PT means a
regular ATTACH; an attached device calling VFIO_DEVICE_ATTACH_IOMMUFD_PT
on the other hand means a REPLACE.

QEMU with this feature should have the vIOMMU maintain a cache of the
guest io page table addresses and assign a unique IOAS to each unique
guest page table.

As the guest writes the page table address to the HW registers qemu should
then use the 'replace domain' operation on VFIO to assign the VFIO device
to the correct de-duplicated page table.

The algorithm where QEMU uses one VFIO container per-device and removes
all the mappings to change the assignment should ideally not be used with
iommufd.

This series is also on Github:
https://github.com/nicolinc/iommufd/commits/wip/replace_ioctl

Thank you
Nicolin Chen