iommu: Add a broken_unmanaged_domain flag in iommu_ops
Some older iommu drivers do not support a fully functional UNMANAGED
iommu_domain. This means that a caller like IOMMUFD will not be able
to work with them.

So, add a "broken_unmanaged_domain" boolean flag in struct iommu_ops,
and set this flag in the following drivers that are obviously lying
about the UNMANAGED domain support:
 * fsl_pamu_domain.c
   Without map/unmap ops in the default_domain_ops, unmanaged domains
   cannot be fully functional in IOMMUFD.
 * mtk_iommu_v1.c
   With a fixed 4M "pagetable", it can only map exactly 4G of memory.
 * tegra-gart.c
   Its notion of attach/detach and groups has to be a complete lie to
   get around all the other API expectations.

Also, set the flag in the following drivers that might work but have
never been tested with VFIO/iommufd:
 * omap-iommu.c
 * msm_iommu.c
 * tegra-smmu.c

Suggested-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>