iommu: Set a boolean no_iommufd in older drivers
To fully work with iommufd, an IOMMU driver needs to support features,
such as fully functional unmanaged iommu_domain, default_domain, and so
on. There are a few older drivers that don't support these so well.

Add a "bool no_iommufd" in struct iommu_ops, and set it to true in all
of these drivers. This allows us to simply return a -EOPNOTSUPP in the
IOMMU core for these drivers, whenever they interact with iommufd.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>