iommufd: Change iommufd_hw_pagetable_has_group to device centric
The iommufd_hw_pagetable_has_group() is a little hack to check whether
a hw_pagetable has an idev->group attached. This isn't device-centric
firstly, but also requires the hw_pagetable to maintain a device list
with a devices_lock, which gets overcomplicated among the routines for
different use cases, upcoming nested hwpts especially.

Since we can compare the iommu_group->domain and the hwpt->domain to
serve the same purpose while being device centric, change it and drop
the device list with the the devices_lock accordingly.

Note that, in the detach routine, it previously checked !has_group as
there was a list_del to the device list, while now it still has the
device being attached to the hwpt, so the if logic gets reversed.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>