iommufd: Drop iommufd_hw_pagetable_has_group()
(This patch should be squashed into the nesting infrastructure series.)

The iommufd_hw_pagetable_has_group() is a little hack to check whether
a hw_pagetable has an idev->group attached. This isn't device-centric
firstly, but also requires the hw_pagetable to maintain a device list
with a devices_lock, which gets overcomplicated among the routines for
different use cases, nesting especially.

Since we can compare group->domain and hwpt->domain to serve the same
purpose of that helper function, drop it and the device list with the
the devices_lock accordingly.

Also, the refcount "devices_users == 1" means the device list is empty.
So the current !iommufd_hw_pagetable_has_group() is always true if the
devices_users condition is met. So simply drop those two lines instead
of doing a domain comparison.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>