iommu/arm-smmu-v3: Add an ARM_SMMU_DOMAIN_ABORT stage
With a 2-stage translation configuration, the stage-1 transaltion can be
controlled by a Guest OS that set its STRTAB_STE_0_CFG field to "abort".
Then, the 2-stage STE, configured by the host kernel, should abort the
entire translation.

Add an ARM_SMMU_DOMAIN_ABORT stage for the nested domain support in the
following patch. And keep master->domain to be NULL so it will go with
the abort setup in arm_smmu_write_strtab_ent() function.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>