iommufd: Simplify the lifecycle of a hwpt
The HWPT is connected to the ioas when it is finished being created and
should remain connected until it is destroyed.

Instead of trying to disconnect the HWPT manually during
iommufd_device_detach() simply use the object reference counting mechanism
and free it when the refcount reaches 0.

This means the ioas->hwpt_list can have an item with a 0 refcount during
races, a prior patch made sure this is safe.

Add a new API iommufd_hw_pagetable_finalize() to formalize this contract.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>