iommufd: Replace the hwpt->devices list with iommufd_group
The devices list was used as a simple way to avoid having per-group
information. Now that this seems to be unavoidable, just commit to
per-group information fully and remove the devices list.

The iommufd_group stores the currently assigned hwpt for the entire group
and we can manage the per-device attach/detach with a simple counter.

For destruction the flow is organized to make the following patches
easier, the actua call to iommufd_object_destroy_user() is done at the top
of the call chain without holding any locks. The hwpt to be destroyed is
returned out from the locked region to make this possible. Later patches
create locking that requires this.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>