iommufd: Replace the hwpt->devices list with iommufd_group
The devices list was uses as a simple way to avoid having per-group
information. Now that this seems to be unavoidable, just commit to per-group
information fully and remove the devices list.

The iommufd_group stores the currently assigned hwpt for the entire group
and we can manage the group attach/detach with a simple counter.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>