cover-letter: Add iommufd physical device operations for replace and alloc hwpt
This is the basic functionality for iommufd to support
iommufd_device_replace() and IOMMU_HWPT_ALLOC.

iommufd_device_replace() allows changing the HWPT associated with the
device to a new IOAS or HWPT. Replace does this in way that failure leaves
things unchanged, and utilizes the iommu iommu_group_replace_domain() API
to allow the iommu driver to perform a non-disruptive change.

IOMMU_HWPT_ALLOC allows HWPTs to be explicitly allocated by the user and
used by attach or replace. At this point it isn't very useful since the
HWPT is the same as the automatically managed HWPT from the IOAS. However
a following series will allow userspace to customize the created HWPT.

The implementation is complicated because we have to introduce some
per-iommu_group memory in iommufd and redo how we think about multi-device
groups to be more explicit. This solves all the locking problems in the
prior attemps.

This series is infrastructure work for the following series which:
 - Add replace for attach
 - Make VFIO use replace
 - Implement driver parameters for domain creation

Once review of this is complete I will keep it on a side branch so the
other series have a stable base. I won't send it to Linus until enough of
the other pieces are completed that this code is used.

This is on github: https://github.com/jgunthorpe/linux/commits/iommufd_hwpt

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>