vfio: Make vfio_device_first_open() to accept NULL iommufd for noiommu
vfio_device_first_open() now covers the below two cases:

1) user uses iommufd (e.g. the group path in iommufd compat mode);
2) user uses container (e.g. the group path in legacy mode);

So caller should either provide a valid iommufd or container. Now caller
can also provide a NULL iommufd pointer for the noiommufd mode in the
group path that is under compat mode. This requires to check if iommufd
is valid first, then check if container is used. If non of them, then it
is noiommu mode.

This also prepares for noiommu support in device cdev path.

Signed-off-by: Yi Liu <yi.l.liu@intel.com>