iommufd: Add IOMMUFD_CMD_SET_IDEV_DATA and IOMMUFD_CMD_UNSET_IDEV_DATA
Add a new pair of ioctls to allow user space to set and unset its iommu-
specific device data at the iommu that a passthrough device is behind.

On platforms with SMMUv3, this new uAPIs will be used to forward a user
space virtual Stream ID of a passthrough device to link to its physical
Stream ID and log into a lookup table, in order for the host kernel to
later run sanity on ATC invalidation requests from the user space, with
ATC_INV commands that have SID fields (virtual Stream IDs).

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>