iommufd: Add mmap infrastructure
For vIOMMU passing through HW resources to user space (VMs), add an mmap
infrastructure to map a region of hardware MMIO pages. The addr and size
should be given previously via a prior IOMMU_VIOMMU_ALLOC ioctl in some
output fields of the structure.

Maintain an mt_mmap per ictx for validations. And give IOMMU drivers a
pair of helpers to add and delete mmap regions onto the mt_mmap.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>