TMP: iommufd: Only enforce IOMMU_RESV_SW_MSI when attaching user-managed HWPT
An IOMMU_RESV_SW_MSI region of a device always needs to be added to a
kernel-managed HWPT, regardless of attaching a device to a user-managed
HWPT or a kernel-managed one, while any other type of reserved regions
should not be added to the kernel-managed HWPT if the attaching HWPT is
a user-managed one in a nesting configuration.

Add a boolean argument to enforce IOMMU_RESV_SW_MSI only, if attaching
HWPT is a user-managed one, i.e. hwpt->parent must be valid, resulting
"!!hwpt->parent" to be true.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>