iommufd: Handle IOMMU_RESV_SW_MSI complication with user-managed HWPT
With one-stage translation, the device attaches to a kernel-managed HWPT
that associates to an IOAS where all the device reserved regions must be
enforced. When the device detaches from one kernel-managed HWPT to another
kernel-managed HWPT, a cleanup of those reserved regions in that IOAS is
needed, unless the HWPT is replaced with another HWPT that shares the same
IOAS. However, user-managed HWPTs introduce more complicated use cases.

With two-stage translation, the device attaches to a user-managed HWPT. In
this case, all the device reserved regions should not be enforced in the
IOAS that its parent HWPT associates to, but IOMMU_RESV_SW_MSI still needs
to be enforced in that IOAS. This requires a new "sw_msi_only" flag to the
iopt_table_enforce_dev_resv_regions().

Meanwhile, when the device detaches from a user-managed HWPT, a cleanup of
all its reserved regions is needed, but some complication happens to cases
with a HWPT replacement. Especially, when switching from a kernel-managed
HWPT to its associating user-managed HWPT, all other reserved regions must
be removed, yet IOMMU_RESV_SW_MSI needs to be kept in the IOAS. This then
requires another "keep_sw_msi" flag in iopt_remove_reserved_iova(), and a
new "sw_msi" tag in struct iopt_reserved, so the "keep_sw_msi" flag would
work.

Lastly, update iommufd_must_enforce_rr() and iommufd_must_remove_rr() to
pass in these new parameters to the iopt functions, so all use cases can
be covered.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>