iommufd: Allow keep IOMMU_RESV_SW_MSI when removing reserved regions
Usually when a device detaches from an attaching HWPT, all of its reserved
regions must be removed from the IOAS. But, when a device switches from a
a kernel-managed HWPT to its associating user-managed HWPT, i.e. a replace,
all other reserved regions must be removed, but IOMMU_RESV_SW_MSI needs to
be kept in the IOAS.

This requires the remover to identify all the IOMMU_RESV_SW_MSI regions in
the ioas->iopt. So, add a new "sw_msi" tag in the iopt_reserved structure
and set it during an iopt_reserve_iova call. Then add a "keep_sw_msi" flag
in iopt_remove_reserved_iova() to exclude IOMMU_RESV_SW_MSI from removing.

Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>