iommu/arm-smmu-v3: Add arm_smmu_domain_needs_cdtab helper
If a master has only a default substream, it can skip CD/translation table
allocations when being attached to an IDENTITY domain, by simply setting
STE to the "bypass" mode (STE.Config[2:0] == 0b100).

If a master has multiple substreams, it will still need a CD table for the
non-default substreams when being attached to an IDENTITY domain, in which
case the STE.Config is set to the "stage-1 translate" mode while STE.S1DSS
field instead is set to the "bypass" mode (STE.S1DSS[1:0] == 0b01).

If a master is attached to a stage-2 domain, it does not need a CD table,
while the STE.Config is set to the "stage-2 translate" mode.

Add arm_smmu_domain_needs_cdtab helper to decide whether the master needs
a CD table or not. Since needing a CD table implies STE.Config will not be
set to bypass, do not always skip the ATS feature detection but check the
return value of the arm_smmu_domain_needs_cdtab helper.

Suggested-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>